feat: collect all the repositories into one
I am trying to recreate the story fo this project. And maybe restore
some of the code that I lost. For now, I have manged to make the
different parts start, though I'm not sure how to connect them together.

Branch 2017 will keep the original state, with minor fixes to give some
sense, probably, as there are divergent branches in one of them (for
exmaple).